requirements_base.txt
Django==1.8.8
django-radiogrid==1.0.4
otree-core[mturk]==1.4.28
